backend/app/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pandas==2.2.3
numpy==1.25.2
scikit-learn==1.4.2
requests==2.34.0fastapi==0.101.1
uvicorn[standard]==0.22.0
pandas==2.2.3
sentence-transformers==2.2.2
faiss-cpu==1.8.0
scikit-learn==1.3.2
python-multipart==0.0.6
beautifulsoup4==4.12.2
requests==2.31.0
tqdm==4.66.1
